requirements.txt
fastapi==0.112.2
langchain==0.3.9
langchain_anthropic==0.2.3
langchain_core==0.3.21
langchain_openai==0.2.3
langchain-unstructured[local]==0.1.6
langsmith==0.1.137
moviepy==2.1.2
openai==1.52.2
opencv_contrib_python==4.10.0.84
opencv_python==4.10.0.84
opencv_python_headless==4.10.0.84
Pillow==10.4.0
pydantic==2.9.2 
pydub==0.25.1
pymongo==4.8.0
python-docx==1.1.2
python-dotenv==1.0.1
python-magic==0.4.27
python-multipart==0.0.9
pytz==2024.1
Requests==2.32.3
scenedetect==0.6.4
sentence_transformers==3.0.1
torch==2.4.1
unstructured==0.15.7
unstructured-client==0.27.0
uvicorn==0.30.6
supabase==2.15.0
httpx==0.27.2


# ==========FramePack lib==========
accelerate==1.6.0
diffusers==0.33.1
transformers==4.46.2
sentencepiece==0.2.0
# pillow==11.1.0
av==12.1.0
numpy==1.26.2
scipy==1.12.0
# requests==2.31.0
torchsde==0.2.6
einops==0.8.1
safetensors==0.5.3
# opencv-contrib-python==4.11.0.86

# ==================================